// src/telegram/enums/scenes.enum.ts
export enum SCENES {
    FIRST_ENTER = 'FIRST_ENTER',
    MAIN_MENU = 'MAIN_MENU',

    // CREATE SCENES
    CREATE_TRAINING = 'CREATE_TRAINING',
    CREATE_EXERCISE = 'CREATE_EXERCISE',

    // START TRAINING SCENES
    WORKOUT_TRAININGS = 'WORKOUT_TRAININGS',
    WORKOUT_EXERCISES = 'WORKOUT_EXERCISES',
    WORKOUT_APPROACHES = 'WORKOUT_APPROACHES',

    // EDIT TRAINING SCENES
    EDIT_MAIN = 'EDIT_MAIN',
    EDIT_TRAINING = 'EDIT_TRAINING',
    EDIT_EXERCISE = 'EDIT_EXERCISE',
}